hdri_api_server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
pillow>=10.0
numpy>=1.24
